Modern C++ M05: Tried header file for custom array class plus non member definitions with custom array, but failed
